+++ app/src/main/java/com/pilot51/voicenotify/ui/theme/Theme.kt
@@ -22,13 +22,7 @@ import androidx.compose.ui.unit.sp
 
 val ClearBlue = Color(0xFF286EFF)
 
-val md_theme_light_primary = Color(0xFF476810)
-val md_theme_light_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_primaryContainer = Color(0xFFC7F089)
 
-val md_theme_dark_primary = Color(0xFFACD370)
-val md_theme_dark_onPrimary = Color(0xFF213600)
-val md_theme_dark_primaryContainer = Color(0xFF324F00)
 
 private val LightColorScheme = lightColorScheme(
     primary = ClearBlue,
@@ -38,15 +32,6 @@ private val LightColorScheme = lightColorScheme(
     primaryContainer = Color(0xFF1E4696),
     surfaceVariant = Color(0xFFf2f3f8),
 
-    /* Other default colors to override
-    background = Color(0xFFFFFBFE),
-    surface = Color(0xFFFFFBFE),
-    onPrimary = Color.White,
-    onSecondary = Color.White,
-    onTertiary = Color.White,
-    onBackground = Color(0xFF1C1B1F),
-    onSurface = Color(0xFF1C1B1F),
-    */
 )
 
 
@@ -55,10 +40,8 @@ private val DarkColorScheme = darkColorScheme(
     secondary = Color(0xFF2A54A5),
     tertiary = Color(0xFF2A54A5),
     // background color
-    background = Color(0xFF010101),
-
+    surface = Color(0xFF010001),
     primaryContainer = Color(0xFF1E4696),
-
     surfaceVariant = Color(0xFF121212),
 )
